Did the ol' switcharoo
