+++ sparse_convolution/__init__.py
@@ -1,4 +1,4 @@
 from sparse_convolution.sparse_convolution import Toeplitz_convolution2d
-from sparse_convolution.minimal_toeplitz import MinimalToeplitzConvolver
+from sparse_convolution.old_version import Toeplitz_convolution2d_Old
 
 __version__ = '0.1.5'
