+++ sparse_convolution/sparse_convolution.py
@@ -1,18 +1,22 @@
-from typing import Tuple, Optional, Union
-
-import scipy.sparse
 import numpy as np
+from typing import Tuple, Optional, Union
+import scipy.sparse
 
 class Toeplitz_convolution2d():
     """
     Convolve a 2D array with a 2D kernel using the Toeplitz matrix
     multiplication method. This class is ideal when 'x' is very sparse
-    (density<0.01), 'x' is small (shape <(1000,1000)), 'k' is small (shape
-    <(100,100)), and the batch size is large (e.g. 1000+). Generally, it is
-    faster than scipy.signal.convolve2d when convolving multiple arrays with the
-    same kernel. It maintains a low memory footprint by storing the toeplitz
-    matrix as a sparse matrix.
+    (density<0.01) and the batch size is large (e.g. 1000+). This class
+    exploits the sparsity of the input matrix/matrices in order to speed
+    up the convolution, so if many matrices are supplied as one batch,
+    and they may be sparse individually but form a dense matrix when overlayed,
+    the speedup gains will diminish. Therefore if using large batches, it's good
+    for the matrices to be correlated and/or extra sparse.
+    Generally, it is faster than scipy.signal.convolve2d when convolving multiple arrays with the
+    same kernel. It maintains a low memory footprint by storing only the
+    necessary columns of the toeplitz matrix as a sparse matrix.
     RH 2022
+    VJ 2024
 
     Attributes:
         x_shape (Tuple[int, int]):
@@ -71,78 +75,31 @@ class Toeplitz_convolution2d():
         assert isinstance(x_shape, (tuple, list)), f"x_shape must be a tuple. Found: {type(x_shape)}"
         assert all([isinstance(s, (int, float, np.integer, np.floating)) for s in x_shape]), f"x_shape must be a tuple of integers. Found: {[type(s) for s in x_shape]}"
         x_shape = (int(x_shape[0]), int(x_shape[1]))
-
         assert isinstance(k, np.ndarray), "k must be a numpy array"
         assert k.ndim == 2, "k must be a 2D array"
-
         assert isinstance(mode, str), "mode must be a string"
         assert mode in ['full', 'same', 'valid'], "mode must be 'full', 'same', or 'valid'"
-
-        # if dtype is not None:
-        #     assert isinstance(dtype, np.dtype), "dtype must be a numpy dtype"
-
-        ## Warn if x_shape is large
-        if verbose > 0:
-            n_nz_elements_expected = x_shape[0]*x_shape[1]*k.shape[0]*k.shape[1]
-            if n_nz_elements_expected >= 1e8:
-                print("Warning: Expected number of non-zero elements in the Toeplitz matrix is large. \n"
-                      f"(x_shape[0]*x_shape[1]*k.shape[0]*k.shape[1]) = {n_nz_elements_expected} non-zero elements. \n"
-                      "This will likely be slow and have a large memory footprint. \n"
-                      "Consider breaking the `x` array into smaller chunks or tiles so that `x_shape` can be smaller and performing the convolution in batches.")
-
-
-        self.k = k = np.flipud(k.copy())
-        self.mode = mode
-        self.x_shape = x_shape
-        dtype = k.dtype if dtype is None else dtype
-
         if mode == 'valid':
             assert x_shape[0] >= k.shape[0] and x_shape[1] >= k.shape[1], "x must be larger than k in both dimensions for mode='valid'"
 
-        self.so = so = size_output_array = ( (k.shape[0] + x_shape[0] -1), (k.shape[1] + x_shape[1] -1))  ## 'size out' is the size of the output array
+        # New assertions
+        assert k.shape[0] > 0 and k.shape[1] > 0, "Kernel must have width and height greater than zero"
+        assert x_shape[0] > 0 and x_shape[1] > 0, "Input matrix must have width and height greater than zero"
 
-        # NOTE: If my analysis is correct, the toeplitz matrices below (made using scipy.sparse.diags)
-        # are stored in either the DIA format, or perhaps CSR or CSC. Either way, it's being stored as
-        # whichever of the sparse formats described here: https://docs.scipy.org/doc/scipy/reference/sparse.html
-        # are determined "appropriate" by the scipy library. The problem is, none of these formats seem
-        # to take advantage of the redundant nature of Toeplitz matrices?? This means that unless the
-        # kernel is largely empty, which it probably isn't, the whole Toeplitz-Block-Toeplitz matrix will
-        # be defined explicitly, which makes working with sizeable input matrices infeasible.
-        # We should be having the Toeplitz matrices defined only by their first column and row, because then
-        # with a 10,000 x 10,000 input matrix and a 100 x 100 kernel, for example, we'd only need to store
-        # ~2,000,000 elements instead of ~10,000,000,000 elements. The memory usage would scale linearly with
-        # the size of the input matrix, instead of quadratically.
-        # TODO: Implement a custom sparse matrix class that takes advantage of the Toeplitz structure,
-        # while still allowing for efficient matrix multiplication. Gotta make sure that this is actually
-        # possible. I think Toeplitz matrices (or perhaps specifically "Circulant" matrices) can be
-        # diagonalized by a Fourier transform, the result of which can then be easily matrix-multiplied
-        # (which corresponds to convolution in the space domain). So, perhaps if we can perform an FFT on
-        # an implicitly stored Toeplitz matrix, then we're laughing. I'd also have to verify that the FFT
-        # method still allows for batching convolution operations, which is a key feature of the current
-        # implementation.
-        # NOTE: After drawing a diagram and understanding the structure of double_toeplitz, I see that
-        # the matrix has a yuge ability to be compressed. The only unique values in the matrix are the
-        # rows of the kernel, which are constantly repeated along the diagonals of sub-tiles of the matrix,
-        # shifted according to the position of the sub-tile. So, basically, the memory usage of the
-        # double_toeplitz can indeed be *massively* reduced. I just need to write a (perhaps second?)
-        # wrapper class for it that can facilitate its matrix multiplication by a vector. Might have to
-        # do this in C.
+        self.x_shape: Tuple[int, int] = x_shape
+        k = np.flipud(k.copy())
+        self.kernel: np.ndarray = k
+        self.mode = mode
+        self.dtype = k.dtype if dtype is None else dtype
+        self.verbose = verbose
 
-        ## make the toeplitz matrices
-        t = toeplitz_matrices = [scipy.sparse.diags(
-            diagonals=np.ones((k.shape[1], x_shape[1]), dtype=dtype) * k_i[::-1][:,None], 
-            offsets=np.arange(-k.shape[1]+1, 1), 
-            shape=(so[1], x_shape[1]),
-            dtype=dtype,
-        ) for k_i in k[::-1]]  ## make the toeplitz matrices for the rows of the kernel
-        tc = toeplitz_concatenated = scipy.sparse.vstack(t + [scipy.sparse.dia_matrix((t[0].shape), dtype=dtype)]*(x_shape[0]-1))  ## add empty matrices to the bottom of the block due to padding, then concatenate
+        # Compute some tings
+        self.padded_kernel_height: int = x_shape[0] + k.shape[0] - 1
+        self.padded_kernel_width: int = x_shape[1] + k.shape[1] - 1
+        self.single_toeplitz_height: int = self.padded_kernel_width
+        self.single_toeplitz_width: int = x_shape[1]
+        self.double_toeplitz_shape: Tuple[int, int] = (self.single_toeplitz_height * self.padded_kernel_height, self.single_toeplitz_width * x_shape[0])
 
-        ## make the double block toeplitz matrix
-        self.dt = double_toeplitz = scipy.sparse.hstack([self._roll_sparse(
-            x=tc, 
-            shift=(ii>0)*ii*(so[1])  ## shift the blocks by the size of the output array
-        ) for ii in range(x_shape[0])]).tocsr()
-    
     def __call__(
         self,
         x: Union[np.ndarray, scipy.sparse.csc_matrix, scipy.sparse.csr_matrix],
@@ -182,63 +139,160 @@ class Toeplitz_convolution2d():
                     * ``batching==False``: Single convolved 2D array of shape
                       *(height, width)*
         """
+        is_sparse = scipy.sparse.issparse(x)
+        if is_sparse:
+            x = x.tocsr()
+        if batching:
+            batch_size = x.shape[0]
+        else:
+            batch_size = 1
+        
         if mode is None:
             mode = self.mode  ## use the mode that was set in the init if not specified
-        issparse = scipy.sparse.issparse(x)
         
-        if batching:
-            x_v = x.T  ## transpose into column vectors
+        # Make B (the flattened and horizontally stacked input matrices)
+        B = x.reshape(batch_size, -1).T
+
+        # Get the indices of empty rows of B
+        if is_sparse:
+            nonzero_B_rows = B.getnnz(axis=1).nonzero()[0]
         else:
-            x_v = x.reshape(-1, 1)  ## reshape 2D array into a column vector
-        
-        if issparse:
-            x_v = x_v.tocsc()
-        
-        out_v = self.dt @ x_v  ## if sparse, then 'out_v' will be a csc matrix
-            
-        ## crop the output to the correct size
+            nonzero_B_rows = np.where(B.any(axis=1))[0]
+
+        ## Warn if x_shape is large
+        density = len(nonzero_B_rows)/B.shape[0]
+        if self.verbose > 0:
+            n_nz_elements_expected = int(density * self.x_shape[0]*self.x_shape[1]*self.kernel.shape[0]*self.kernel.shape[1])
+            if n_nz_elements_expected >= 1e8:
+                print("Warning: The number of non-zero elements in the Toeplitz matrix is large. \n"
+                      f"(d * x_shape[0]*x_shape[1]*k.shape[0]*k.shape[1]) = {n_nz_elements_expected} non-zero elements, \n"
+                      f"where d is the effective density of the input matrix batch (d = {density}). \n"
+                      "This will likely be slow and have a large memory footprint. \n"
+                      "Consider breaking the `x` array into smaller chunks or tiles so that `x_shape` can be smaller and performing the convolution in batches.")
+            if density >= 0.05:
+                print(f"Warning: The density of the input matrix (or the effective density of the batch) is high: density = {density} >= 0.05\n"
+                      "Regular convolution methods may perform better in this case.")
+
+        # Form the bare minimum double Toeplitz matrix
+        cols = nonzero_B_rows
+        rows = self._get_nonzero_rows(cols)
+        data = self._get_values(rows, cols)
+        rows = rows.flatten()
+        rows_per_col = self.kernel.size
+        cols = np.repeat(cols, rows_per_col)
+        DT = scipy.sparse.csr_matrix((data, (rows, cols)), shape=self.double_toeplitz_shape) # TODO: add arg dtype=self.dtype (doing later in case it breaks something)
+
+        # Do the roar
+        out_uncropped = DT @ B
+
+        # Change dtype if necessary
+        if out_uncropped.dtype != self.dtype:
+            out_uncropped = out_uncropped.astype(self.dtype)
+
+        # Unvectorize output
+        so = size_output_array = ((self.x_shape[0] + self.kernel.shape[0] - 1), (self.kernel.shape[1] + self.x_shape[1] -1))
+        # Reconcile it with the original implementation's output shape
+
+        # Crop the output to the correct size
         if mode == 'full':
             t = 0
-            b = self.so[0]+1
+            b = so[0]+1
             l = 0
-            r = self.so[1]+1
+            r = so[1]+1
         if mode == 'same':
-            t = (self.k.shape[0]-1)//2
-            b = -(self.k.shape[0]-1)//2
-            l = (self.k.shape[1]-1)//2
-            r = -(self.k.shape[1]-1)//2
+            t = (self.kernel.shape[0]-1)//2
+            b = -(self.kernel.shape[0]-1)//2
+            l = (self.kernel.shape[1]-1)//2
+            r = -(self.kernel.shape[1]-1)//2
 
             b = self.x_shape[0]+1 if b==0 else b
             r = self.x_shape[1]+1 if r==0 else r
         if mode == 'valid':
-            t = (self.k.shape[0]-1)
-            b = -(self.k.shape[0]-1)
-            l = (self.k.shape[1]-1)
-            r = -(self.k.shape[1]-1)
+            t = (self.kernel.shape[0]-1)
+            l = (self.kernel.shape[1]-1)
+            b = -(self.kernel.shape[0]-1)
+            r = -(self.kernel.shape[1]-1)
 
             b = self.x_shape[0]+1 if b==0 else b
             r = self.x_shape[1]+1 if r==0 else r
-        
+
         if batching:
-            idx_crop = np.zeros((self.so), dtype=np.bool_)
+            idx_crop = np.zeros(so, dtype=np.bool_)
             idx_crop[t:b, l:r] = True
             idx_crop = idx_crop.reshape(-1)
-            out = out_v[idx_crop,:].T
+            out = out_uncropped[idx_crop,:].T
         else:
-            if issparse:
-                out = out_v.reshape((self.so)).tocsc()[t:b, l:r]
+            if is_sparse:
+                out = out_uncropped.reshape(so).tocsc()[t:b, l:r]
             else:
-                out = out_v.reshape((self.so))[t:b, l:r]  ## reshape back into 2D array and crop
+                out = out_uncropped.reshape(so)[t:b, l:r]  ## reshape back into 2D array and crop
         return out
     
-    def _roll_sparse(
-        self,
-        x: scipy.sparse.csr_matrix,
-        shift: int,
-    ):
+    def _get_values(self, row_matrix: np.ndarray, col_vector: np.ndarray) -> np.ndarray:
         """
-        Roll columns of a sparse matrix.
+        Compute the values of the matrix at position (row, col) dynamically.
+        If any (row,col) corresponding to a zero value in the matrix is
+        provided, an exception will be thrown. This is because the
+        current program is already designed to only provide coords of
+        non-zero values, and we can just save some miniscule amount of
+        time by not bothering to make it work for zero values. Or maybe
+        it could be done, I unno, I don't think it matters.
+
+        Parameters:
+        - row_matrix (np.ndarray): The input row matrix.
+        - col_vector (np.ndarray): The input column vector.
+
+        Returns:
+        - np.ndarray: The computed values of the matrix at position (row, col).
+
+        Note:
+        - The input arrays `row_matrix` and `col_vector` should have compatible shapes.
+        - The shape of `row_matrix` should be (C,) or (C, R), where N is the number of columns and K is the number of rows.
+        - The shape of `col_vector` should be (C,) where C is the number of columns.
+        - The returned array will be a (C*R, 2) array
         """
-        out = x.copy()
-        out.row += shift
-        return out
+        # Find which inner Toeplitz this row/col is in,
+        # whilst simultaneously retrieiving its toeplitz-relative row/col
+        tb_col_vector, t_col_vector = np.divmod(col_vector, self.single_toeplitz_width)
+        tb_row_matrix, t_row_matrix = np.divmod(row_matrix, self.single_toeplitz_height)
+
+        # Coordinates of kernel that correspond to the coordinates in row_matrix and col_vector
+        kernel_row_matrix = (self.kernel.shape[0] - 1) - (tb_row_matrix - tb_col_vector[:, None])
+         # Free up memory along the way
+        del tb_col_vector
+        del tb_row_matrix
+        padded_kernel_col_matrix = t_row_matrix - t_col_vector[:, None]
+        del t_col_vector
+        del t_row_matrix
+        padded_kernel_indices = np.stack((kernel_row_matrix, padded_kernel_col_matrix), axis=-1).reshape(-1, 2)
+        del padded_kernel_col_matrix
+        del kernel_row_matrix
+
+        # Return a 1D numpy array of the values of kernel that correspond to the coordinates in padded_kernel_indices
+        result = self.kernel[tuple(padded_kernel_indices.T)]
+
+        return result
+
+    def _get_nonzero_rows(self, col_vector: np.ndarray) -> np.ndarray:
+        """
+        Get the indices of nonzero elements in the columns `col_vector` of the matrix.
+        """      
+        assert col_vector.ndim == 1
+        if col_vector.size == 0:
+            return np.array([], dtype=int)
+        
+        # Compute toeplitz block and relative indices
+        tb_col_vector, t_col_vector = np.divmod(col_vector, self.single_toeplitz_width)
+
+        # Compute row offsets for each column
+        i_range = self.kernel.shape[0] * self.single_toeplitz_height
+        i_offsets_slice = np.arange(self.kernel.shape[1]) + np.arange(0, i_range, self.single_toeplitz_height)[:, None]
+        i_offsets_slice = i_offsets_slice.ravel()  # Flatten for efficient broadcasting
+
+        # Compute full row indices
+        i_matrix = t_col_vector[:, None] + i_offsets_slice
+        i_matrix += tb_col_vector[:, None] * self.single_toeplitz_height
+
+        return i_matrix
+
+
